Move firebase repos to their own packages
