+++ app/src/main/java/dev/android/peru/data/repositories/MeetupsRepo.kt
@@ -1,0 +1,7 @@
+package dev.android.peru.data.repositories
+
+import peru.android.dev.datamodel.Meetup
+
+interface MeetupsRepo {
+    suspend fun getMeetups(): List<Meetup>
+}
