+++ app/src/main/java/dev/android/peru/data/repositories/QuestionnaireRepo.kt
@@ -1,0 +1,7 @@
+package dev.android.peru.data.repositories
+
+import peru.android.dev.datamodel.Questionnaire
+
+interface QuestionnaireRepo {
+    suspend fun getQuestionnaireForMeetup(meetupId: String): Questionnaire?
+}
