+++ app/src/main/java/dev/android/peru/data/source/firebase/repositories/FirebaseMeetupsRepo.kt
@@ -1,16 +1,11 @@
-package dev.android.peru.data
+package dev.android.peru.data.source.firebase.repositories
 
-import dev.android.peru.data.firebase.FirebaseParser
-import dev.android.peru.data.firebase.FirebaseRepo
+import dev.android.peru.data.repositories.MeetupsRepo
+import dev.android.peru.data.source.firebase.FirebaseParser
+import dev.android.peru.data.source.firebase.FirebaseRepo
 import peru.android.dev.datamodel.Meetup
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
-
-interface MeetupsRepo {
-
-    suspend fun getMeetups(): List<Meetup>
-
-}
 
 class FirebaseMeetupsRepo(): MeetupsRepo, FirebaseRepo {
