+++ app/src/main/java/dev/android/peru/data/source/firebase/repositories/FirebaseQuestionnaireRepo.kt
@@ -1,14 +1,12 @@
-package dev.android.peru.data
+package dev.android.peru.data.source.firebase.repositories
 
-import dev.android.peru.data.firebase.FirebaseParser
-import dev.android.peru.data.firebase.FirebaseRepo
+import dev.android.peru.data.repositories.QuestionnaireRepo
+import dev.android.peru.data.source.firebase.FirebaseParser
+import dev.android.peru.data.source.firebase.FirebaseRepo
 import peru.android.dev.datamodel.Questionnaire
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
-interface QuestionnaireRepo {
-    suspend fun getQuestionnaireForMeetup(meetupId: String): Questionnaire?
-}
 
 class FirebaseQuestionnaireRepo(): QuestionnaireRepo, FirebaseRepo {
